Minimal modifications to Subthresh_GIF to fit subthreshold training sets without crashing
Adapt GIF.fitVoltageReset to leave GIF spike-related parameters unchanged if no spks are found in the training set. Hack GIF.fitSubthresholdDynamics to run without crashing if no spks are found: GIF.fitSubthresholdDynamicsd uses GIF.simulate to evaluate the var_explained on V, but GIF.simulate crashes if a spike-triggered current is not provided; creating a dummy spike-triggered current in GIF.fitSubthresh... solves this problem. Modify GIF.fitSubthresholdDynamics_Build_Xmatrix_Yvector to build an X-matrix that does not include columns for the spike-triggered current if no spks are present in the training set.