Create KappaComparator class to hold methods for batch extraction and plotting of membrane filters
Create KappaComparator class with methods for batch simulation and voltage-dependent PSD extraction based on the 'Extract detailed power spectrum density' block (and the next few blocks) in SubthresholdGIFAnalysis.py. Use this class to investigate the role of IA and KSlow in shaping the membrane filter. Note that as of this commit, SubthresholdGIFAnalysis.py has not been refactored to make use of the new KappaComparator class (just in case...).