Fix broadcasting error in run_simulation.py
GIFnet_pipeline/run_simulation.py attempts to broadcast 5HT and GABA inputs to
match the number of cells in the GIFnet to be simulated. This causes a
broadcasting error if input for more than one cell is provided but the number
of cells in the GIFnet is zero. This patch introduces a partial fix by skipping
input broadcasting when the `--no-ser/--no-gaba` flag is supplied.