Numerically estimate resting membrane potential
In GIF-like models with more than one conductance (ie, the AugmentedGIF)
the resting membrane potential does not exactly match the reversal of
the leak conductance E_l. This patch adds a function for numerically
estimating the resting membrane potential of GIF-like models using the
`simulateDeterministic_forceSpikes` method with an empty set of spike
times.