Add accelerated method for computing PSTH
Computing the PSTH of a spiketrain can take a long time because
convolution is slow. This patch implements a new method for computing
the PSTH that takes advantage of the sparsity of spiketrains. The new
method is faster for spiketrains that are >65% sparse, and is
selected automatically when it is expected to provide a speedup. For
very sparse spiketrains, the new method is ~30X faster.

Summary of changes
------------------

- Implemented sparse convolution method as Spiketrain._sparse_PSTH
- Redefined Spiketrain.PSTH to call either _sparse_PSTH or
  _dense_PSTH (old method) depending on spiketrain sparsity
- Added unit tests to ensure _sparse_PSTH and _dense_PSTH produce
  identical output, up to numerical precision