Fix max-current to conductance calculation for I_A
ser_currents_<temperature>.ipynb notebooks calculated the maximal
conductance of I_A based on the amplitude of the associated current, but
results were incorrect because the wrong driving force was used. This
patch corrects the mistake by basing the driving force on the
equilibrium potential for potassium rather than resting membrane
potential.