Fix IAmod gating functions
The toy model of a neuron with IA (IAmod) reimplements the m and h
equilibrium gating functions inside of its simulation method for
performance reasons, but before this commit they were reimplemented
incorrectly. Specifically, the scaling parameter A was omitted and the
value 1 was used instead. So far, this has had very little effect
because the values of A being used were close to 1, but would
significantly change the behaviour of the model if A were far from 1.